+++ NEWS.md
@@ -4,6 +4,7 @@
 * Updated `tidy.corpus` and `glance.corpus` for changes to quanteda API 
 * Removed the deprecated `pair_count` function, which is now in the in-development widyr package
 * Added tidiers for LDA models from the `mallet` package
+* Added the Loughran and McDonald dictionary of sentiment words specific to financial reports
 
 # tidytext 0.1.2